Funcionarios + Script SQL

+++ src/main/java/com/github/joao/model/Funcionario.java
@@ -1,13 +1,15 @@
 package com.github.joao.model;
 
+import java.math.BigDecimal;
+
 public class Funcionario extends Pessoa {
     private int horasTrabalho;
-    private double salario;
+    private BigDecimal salario;
 
     public Funcionario() {
     }
 
-    public Funcionario(int id, String nome, String cpf, String telefone, String endereco, int horasTrabalho, double salario) {
+    public Funcionario(int id, String nome, String cpf, String telefone, String endereco, int horasTrabalho, BigDecimal salario) {
         super(id, nome, cpf, telefone, endereco);
         this.horasTrabalho = horasTrabalho;
         this.salario = salario;
@@ -17,7 +19,7 @@ public class Funcionario extends Pessoa {
         return horasTrabalho;
     }
 
-    public double getSalario() {
+    public BigDecimal getSalario() {
         return salario;
     }
 
@@ -28,8 +30,8 @@ public class Funcionario extends Pessoa {
             throw new IllegalArgumentException("O funcionário pode trabalhar no máximo 10 horas diárias e este dado é obrigatório!");
     }
 
-    public void setSalario(double salario) {
-        if (salario > 0 && salario < 100000.00)
+    public void setSalario(BigDecimal salario) {
+        if (salario.doubleValue() > 0 && salario.doubleValue() < 100000.00)
             this.salario = salario;
         else throw new IllegalArgumentException("Salário deve ser um valor entre R$0 e 100000.00 e não pode ser nulo!");
     }
